requirements.txt
pillow
pytube
youtube-dl
yt-dlp
moviepy
requests
SpeechRecognition
opencv-python
imageio
numpy
ffmpeg
ffprobe
edge-tts
gTTS
pyttsx3
Flask
werkzeug